Rename signing validation to witnessing validation
