+++ test/integration/ledger/node/tx.js
@@ -4,8 +4,8 @@ const { decodeTx } = require('cardano-hw-interop-lib')
 const { LedgerCryptoProvider } = require('../../../../src/crypto-providers/ledgerCryptoProvider')
 const { NETWORKS } = require('../../../../src/constants')
 const { determineSigningMode, getTxBodyHash } = require('../../../../src/crypto-providers/util')
-const { validateWitnessing } = require('../../../../src/crypto-providers/signingValidation')
-const { validateTxBeforeSigning } = require('../../../../src/transaction/transactionValidation')
+const { validateWitnessing } = require('../../../../src/crypto-providers/witnessingValidation')
+const { validateTxBeforeWitnessing } = require('../../../../src/transaction/transactionValidation')
 
 const { signingFiles } = require('./signingFiles')
 const { getTransport } = require('./speculos')
@@ -893,7 +893,7 @@ const transactions = {
 }
 
 async function testTxWitnessing(cryptoProvider, transaction) {
-  validateTxBeforeSigning(transaction.cborHex)
+  validateTxBeforeWitnessing(transaction.cborHex)
   const txCbor = Buffer.from(transaction.cborHex, 'hex')
   const tx = decodeTx(txCbor)
